Created DAG class, added code to LCA.java
The code added to LCA.java is the beginning of an implementation to find the Lowest Common Ancestor of two nodes in a Directed Acyclic Graph. My chosen method of implementing a solution to this problem is to use ArrayLists of type DAGNode to store information regarding the parent and child nodes of the DAGNodes. These lists can then be iterated through to find the lowest common node in both, which would indicate the lowest common ancestor of the two DAGNodes in question.